control deletion of temporary generation artifacts by the controller log level (DEBUG -> don't delete)

+++ trunk/olympos/chronos/scr/applications/requirements/application/include/controller/class.ExpeditorGeneratorController.php
@@ -188,9 +188,10 @@ class ExpeditorGeneratorController extends BatchController
 		$session = &SessionData::getInstance();
 	
 		// cleanup
-		unlink($session->get($this->TEMP_UWM_EXPORT_PATH));
-		unlink($session->get($this->TEMP_PROPERTIES_PATH));
-
+		if (!Log::isDebugEnabled(__CLASS__)) {
+			unlink($session->get($this->TEMP_UWM_EXPORT_PATH));
+			unlink($session->get($this->TEMP_PROPERTIES_PATH));
+		}
 	}
 	// PROTECTED REGION END
