+++ trunk/olympos/chronos/scr/applications/requirements/application/include/controller/class.GenerateCodeController.php
@@ -213,9 +213,10 @@ class GenerateCodeController extends BatchController
 		$session = &SessionData::getInstance();
 
 		// cleanup
-		unlink($session->get(self::TEMP_UWM_EXPORT_PATH));
-		unlink($session->get(self::TEMP_PROPERTIES_PATH));
-
+		if (!Log::isDebugEnabled(__CLASS__)) {
+			unlink($session->get(self::TEMP_UWM_EXPORT_PATH));
+			unlink($session->get(self::TEMP_PROPERTIES_PATH));
+		}
 	}
 	// PROTECTED REGION END
