+++ trunk/olympos/chronos/scr/applications/requirements/application/include/controller/class.UWMExporterController.php
@@ -229,12 +229,13 @@ class UWMExporterController extends BatchController
 		$this->check('File written to output');
 
 		// cleanup
-		unlink($session->get($this->TEMP_UWM_EXPORT_PATH));
-		unlink($session->get($this->TEMP_PROPERTIES_PATH));
-		unlink("$tmpUmlExportPath/uml-generated.uml");
-		unlink($exportFile);
-		rmdir($session->get($this->TEMP_UML_EXPORT_PATH));
-
+		if (!Log::isDebugEnabled(__CLASS__)) {
+			unlink($session->get($this->TEMP_UWM_EXPORT_PATH));
+			unlink($session->get($this->TEMP_PROPERTIES_PATH));
+			unlink("$tmpUmlExportPath/uml-generated.uml");
+			unlink($exportFile);
+			rmdir($session->get($this->TEMP_UML_EXPORT_PATH));
+		}
 		ExportShutdownHandler::success();
 	}
 	// PROTECTED REGION END
